make a general ecmwf dataset and inherit for era5(land)
